[CHORE] #35 : SignInBottomSheet -> SignInScreen, SignInScreen -> SignInContent
